Change string to DU
When viewing this sample, I thought it might be nice to remove the "_" match on (old) line 176.  To do so, I thought it might be nicer to use a discriminated union instead of strings to indicate state.  Hence this PR -- I hope you like it :)  All comments very much welcome -- I am a distinct newbie.